Cast the input to isalpha et al to integer
* libguile/gc_os_dep.c (GC_linux_stack_base) [LINUX_STACKBOTTOM]: cast
  input of ctype functions to int

* libguile/inet_aton.c (inet_aton): cast input of ctype functions to int

* libguile/read.c (scm_scan_for_encoding): cast input of isalnum to int

* libguile/win32-socket.c (scm_i_socket_uncomment): cast input of isspace
  to int

+++ libguile/gc_os_dep.c
@@ -1867,11 +1867,14 @@ void *scm_get_stack_base()
     /* Skip the required number of fields.  This number is hopefully	*/
     /* constant across all Linux implementations.			*/
       for (i = 0; i < STAT_SKIP; ++i) {
-	while (isspace(c)) c = stat_buf[buf_offset++];
-	while (!isspace(c)) c = stat_buf[buf_offset++];
+	while (isspace ((int) c)) 
+          c = stat_buf[buf_offset++];
+	while (!isspace ((int) c)) 
+          c = stat_buf[buf_offset++];
       }
-    while (isspace(c)) c = stat_buf[buf_offset++];
-    while (isdigit(c)) {
+      while (isspace ((int) c)) 
+        c = stat_buf[buf_offset++];
+    while (isdigit ((int) c)) {
       result *= 10;
       result += c - '0';
       c = stat_buf[buf_offset++];
